heart_mlops/requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
joblib>=1.3
pyyaml>=6.0
dvc>=3.50
dvc-gdrive>=3.0 ; platform_system != 'Windows'  # optional remote example
pytest>=7.4
flake8>=7.0
black>=24.0
isort>=5.13
